EA-710 got rid of the frontend and restructured the project
