--- templates/about.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>THR - About</title>
-</head>
-<body>
-    <h1>About Page</h1>
-</body>
-</html>
