--- templates/home.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>THR - Home</title>
-</head>
-<body>
-    <h1>The Track Hub Registry</h1>
-    <h3>A global centralised collection of publicly accessible track hubs</h3>
-    <div>
-        {% if user.is_authenticated %}
-            <a href="{% url 'logout' %}">Logout</a>
-        {% else %}
-            <a href="{% url 'login' %}">Login</a>
-        {% endif %}
-    </div>
-
-</body>
-</html>
