--- templates/registration/login.html
@@ -1,16 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>Login</h2>
-
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <input type="submit" value="Login">
-</form>
-
-<a href="{% url 'thr_home' %}">Back to Home</a>
-<a href="{% url 'password_reset' %}">Reset password</a>
-<a href="{% url 'register' %}">Register</a>
-
-{% endblock %}
