--- templates/registration/password_change_done.html
@@ -1,7 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>Password changed</h2>
-
-<a href="{% url 'dashboard' %}">Back to dashboard</a>
-{% endblock %}
