--- templates/registration/password_reset_complete.html
@@ -1,7 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>Password reset complete</h2>
-
-<a href="{% url 'login' %}">Back to login</a>
-{% endblock %}
