--- templates/registration/password_reset_confirm.html
@@ -1,11 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>Confirm password reset</h2>
-
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <input type="submit" value="Confirm">
-</form>
-{% endblock %}
