--- templates/registration/password_reset_email.html
@@ -1,3 +1,0 @@
-Someone asked for password reset for email {{ email }}.
-Follow the link below:
-{{ protocol}}://{{ domain }}{% url 'password_reset_confirm' uidb64=uid token=token %}
