--- templates/registration/password_reset_form.html
@@ -1,13 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>Send password reset link</h2>
-
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <input type="submit" value="Reset">
-</form>
-
-<a href="{% url 'dashboard' %}">Back to dashboard</a>
-{% endblock %}
