--- templates/registration/password_reset_subject.txt
@@ -1,1 +1,0 @@
-Reset password
