--- templates/user/dashboard.html
@@ -1,15 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-    <h1>{{ user.username|title }}'s dashboard</h1>
-
-
-    <div>
-        {% if user.is_authenticated %}
-            <a href="{% url 'logout' %}">Logout</a>
-            <a href="{% url 'password_change' %}">Change password</a>
-        {% else %}
-            <a href="{% url 'login' %}">Login</a>
-        {% endif %}
-    </div>
-{% endblock %}
