--- templates/user/register.html
@@ -1,13 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>Register</h2>
-
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <input type="submit" value="Register">
-</form>
-
-<a href="{% url 'login' %}">Back to login</a>
-{% endblock %}
