+++ trackhubs/urls.py
@@ -11,3 +11,12 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 """
+
+from django.urls import path
+
+from trackhubs import views
+
+urlpatterns = [
+    path('', views.TrackHubList.as_view()),
+    path('<int:pk>/', views.TrackHubDetail.as_view()),
+]
