+++ users/views.py
@@ -12,19 +12,65 @@
    limitations under the License.
 """
 
-from django.contrib.messages.views import SuccessMessageMixin
-from django.urls import reverse_lazy
-from django.views.generic import CreateView, TemplateView
+from django.contrib.auth import logout
+from rest_framework import status, authentication, permissions
+from rest_framework.authtoken.models import Token
+from rest_framework.response import Response
+from rest_framework.views import APIView
 
-from .forms import CustomUserCreationForm
-
-
-class DashboardView(TemplateView):
-    template_name = 'user/dashboard.html'
+from users.serializers import RegistrationSerializer
 
 
-class RegistrationView(SuccessMessageMixin, CreateView):
-    template_name = 'user/register.html'
-    success_url = reverse_lazy('login')
-    form_class = CustomUserCreationForm
-    success_message = "Your profile was created successfully"
+class RegistrationViewAPI(APIView):
+    """
+    User registration endpoint, if the request is successful,
+    an HttpResponse is returned with the access token
+    :param request: the request
+    :returns: the data if the request was successful otherwise it return an error message
+    """
+
+    def post(self, request):
+        serializer = RegistrationSerializer(data=request.data)
+        data = {}
+        if serializer.is_valid():
+            new_user = serializer.save()
+            data['response'] = 'User registered successfully!'
+            token = Token.objects.create(user=new_user).key
+            data['token'] = token
+            return Response(data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class LogoutViewAPI(APIView):
+    """
+    Log the users out if they are already logged in,
+    and delete the access token from the database
+    """
+    authentication_classes = [authentication.TokenAuthentication]
+    permission_classes = [permissions.IsAuthenticated]
+
+    def post(self, request):
+        """
+        Returns the response message 200 or 401 (Invalid token)
+        """
+        request.user.auth_token.delete()
+        logout(request)
+        return Response({"success": "Successfully logged out."}, status.HTTP_200_OK)
+
+
+class UserDetailsView(APIView):
+    """
+    Get the user details when providing a valid token
+    """
+    authentication_classes = [authentication.TokenAuthentication]
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get(self, request):
+        user = request.user
+        return Response({
+            'username': user.username,
+            'email': user.email,
+            'first_name': user.first_name,
+            'last_name': user.last_name
+        })
